Программа на Java, реализация нормализованного кол-ва элементов, вариант 3
